feat: 모임 상세 페이지 친구 상세 정보 API 연동 및 UI 개선

+++ frontend/app/src/main/java/com/markoala/tomoandroid/data/model/friends/FriendModels.kt
@@ -3,7 +3,7 @@ package com.markoala.tomoandroid.data.model.friends
 data class FriendProfile(
     val username: String,
     val email: String,
-    val friendship: Double,
+    val friendship: Int,
     val createdAt: String
 )
